+++ internal/provider/resources/resource_datapower_mgmt_interface_test.go
@@ -38,7 +38,6 @@ func TestAccResourceMgmtInterface(t *testing.T) {
 		Check: resource.ComposeTestCheckFunc([]resource.TestCheckFunc{
 			resource.TestCheckResourceAttr("datapower_mgmt_interface.test", "enabled", "false"),
 			resource.TestCheckResourceAttr("datapower_mgmt_interface.test", "local_port", "5550"),
-			resource.TestCheckResourceAttr("datapower_mgmt_interface.test", "acl", "xml-mgmt"),
 			resource.TestCheckResourceAttr("datapower_mgmt_interface.test", "slm_peering", "10"),
 			resource.TestCheckResourceAttr("datapower_mgmt_interface.test", "ssl_config_type", "server"),
 			resource.TestCheckResourceAttr("datapower_mgmt_interface.test", "local_address", "0.0.0.0"),