+++ docs/guides/changelog.md
@@ -13,6 +13,7 @@ description: |-
 - BREAKING CHANGE: Removed fixed setting `AAAPolicy` from `InteropService`
 - BREAKING CHANGE: Removed fixed setting `ACL` from `SSHService`
 - BREAKING CHANGE: Removed fixed setting `ACL` from `WebGUI`
+- BREAKING CHANGE: Removed fixed setting `ACL` from `MgmtInterface`
 
 ## v0.15.0